feat(filterql-jpa): add reducer support for computed fields with SQL aggregation
- Add AggregateQueryExecutor for batch aggregate queries (SUM, COUNT, AVG, MIN, MAX)
- Add prefetchAggregates method to MultiQueryFetchStrategy for batch execution
- Modify applyComputedFieldsToRow to use pre-fetched aggregate values
- Add ReducerMapping support to ComputedFieldInfo with isAggregate() method
- Update CompanyDto and CompanyComputedProvider to use reducers
- Update UltimateNestedBenchmarkTest to use InstanceResolver

Enables SQL-side aggregation for computed fields that depend on collection paths.
Instead of passing entire collections to providers, reducer values (Long) are passed.

+++ adapters/java/filterql-jpa/src/test/java/io/github/cyfko/filterql/jpa/entities/projection/_4/Department.java
@@ -18,7 +18,7 @@ public class Department {
 
     private String name;
 
-    private int budget;
+    private Long budget;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "company_id")
@@ -30,7 +30,7 @@ public class Department {
     public Department() {
     }
 
-    public Department(String name, int budget) {
+    public Department(String name, Long budget) {
         this.name = name;
         this.budget = budget;
     }
@@ -51,11 +51,11 @@ public class Department {
         this.name = name;
     }
 
-    public int getBudget() {
+    public Long getBudget() {
         return budget;
     }
 
-    public void setBudget(int budget) {
+    public void setBudget(Long budget) {
         this.budget = budget;
     }
